// security/front/src/app/models/registrationRequest.model.ts
export interface RegistrationRequest {
  id: number;
  name: string;
  surname: string;
  username: string;
  password: string;
  city: string;
  country: string;
  street: string;
  streetNumber: string;
  phone: string;
  pib: string;
  packageType: string;
  type: string;
  email: string;
  requestStatus: string;
}